refactor: переписаны модели
