+++ statapp/models/utils.py
@@ -1,0 +1,2 @@
+def yx_header(count):
+    return ['Y'] + [f'X{i}' for i in range(1, count)]
